--- user/src/main/java/user/common/config/entrypoint/CustomAuthenticationEntryPoint.java
@@ -1,32 +1,0 @@
-package user.common.config.entrypoint;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import global.api.Api;
-import global.errorcode.ErrorCode;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.OutputStream;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-        AuthenticationException authException) throws IOException, ServletException {
-
-        Api<Object> body = Api.ERROR(ErrorCode.MISSING_REQUIRED_HEADER);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        OutputStream responseStream = response.getOutputStream();
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.writeValue(responseStream, body);
-        responseStream.flush();
-
-    }
-}
