--- user/src/main/java/user/controller/model/duplication/DuplicationNameRequest.java
@@ -1,14 +1,0 @@
-package user.controller.model.duplication;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class DuplicationNameRequest {
-
-    private String name;
-
-}
