--- user/src/main/java/user/controller/model/login/UserLoginRequest.java
@@ -1,16 +1,0 @@
-package user.controller.model.login;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserLoginRequest {
-
-    private String email;
-
-    private String password;
-
-}
