+++ user/src/main/java/user/domain/jwt/model/TokenDto.java
@@ -1,0 +1,19 @@
+package user.domain.jwt.model;
+
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TokenDto {
+
+    private String token;
+
+    private LocalDateTime expiredAt;
+
+}
