+++ user/src/main/java/user/domain/jwt/model/TokenResponse.java
@@ -1,0 +1,23 @@
+package user.domain.jwt.model;
+
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenResponse {
+
+    private String accessToken;
+
+    private LocalDateTime accessTokenExpiredAt;
+
+    private String refreshToken;
+
+    private LocalDateTime refreshTokenExpiredAt;
+
+}
