+++ user/src/main/java/user/domain/jwt/model/TokenValidationRequest.java
@@ -1,0 +1,14 @@
+package user.domain.jwt.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TokenValidationRequest {
+
+    private TokenDto tokenDto;
+
+}
