+++ user/src/main/java/user/domain/jwt/model/TokenValidationResponse.java
@@ -1,0 +1,16 @@
+package user.domain.jwt.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TokenValidationResponse {
+
+    private Long userId;
+
+}
