+++ user/src/main/java/user/domain/user/controller/model/User.java
@@ -1,0 +1,21 @@
+package user.domain.user.controller.model;
+
+import db.domain.user.enums.UserRole;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Getter
+@EqualsAndHashCode
+public class User {
+
+    private Long id;
+    private String email;
+    private UserRole role;
+
+}
