+++ user/src/main/java/user/domain/user/controller/model/duplication/DuplicationNameRequest.java
@@ -1,0 +1,14 @@
+package user.domain.user.controller.model.duplication;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class DuplicationNameRequest {
+
+    private String name;
+
+}
