--- user/src/main/java/user/security/jwt/filter/JwtAuthFilter.java
@@ -1,53 +1,0 @@
-package user.security.jwt.filter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.filter.OncePerRequestFilter;
-import user.security.jwt.service.TokenService;
-import user.security.service.AuthorizationService;
-
-@Slf4j
-@RequiredArgsConstructor
-public class JwtAuthFilter extends OncePerRequestFilter {
-
-    private final AuthorizationService authorizationService;
-    private final TokenService tokenService;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-        FilterChain filterChain) throws ServletException, IOException {
-
-        String authorizationHeader = request.getHeader("Authorization");
-
-        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
-            String token = authorizationHeader.substring(7);
-
-            Long userId = tokenService.validationToken(token);
-
-            // token 검증에서 예외를 발생시키므로 불필요한 로직
-//            if (userId == null) {
-//                throw new TokenException(UserErrorCode.USER_NOT_FOUND);
-//            }
-
-            UserDetails userDetails = authorizationService.loadUserByUsername(
-                Long.toString(userId));
-
-            // UserDetails, Password,  Role -> 접근 권한 인증 Token 생성
-            UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(
-                userDetails, null, userDetails.getAuthorities());
-
-            // 현재 Request 의 Security Context 에 접근 권한 설정
-            SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
-        }
-
-        filterChain.doFilter(request, response); // 다음 필터로 전달
-    }
-}
