--- user/src/main/java/user/security/jwt/ifs/TokenHelperIfs.java
@@ -1,12 +1,0 @@
-package user.security.jwt.ifs;
-
-import java.util.Map;
-import user.security.jwt.model.TokenDto;
-
-public interface TokenHelperIfs {
-
-    TokenDto issueAccessToken(Map<String, Object> data);
-    TokenDto issueRefresh(Map<String, Object> data);
-    Map<String, Object> validationTokenWithThrow(String token);
-
-}
