--- user/src/main/java/user/security/jwt/model/TokenDto.java
@@ -1,19 +1,0 @@
-package user.security.jwt.model;
-
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TokenDto {
-
-    private String token;
-
-    private LocalDateTime expiredAt;
-
-}
