--- user/src/main/java/user/security/jwt/model/TokenResponse.java
@@ -1,23 +1,0 @@
-package user.security.jwt.model;
-
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenResponse {
-
-    private String accessToken;
-
-    private LocalDateTime accessTokenExpiredAt;
-
-    private String refreshToken;
-
-    private LocalDateTime refreshTokenExpiredAt;
-
-}
