--- user/src/main/java/user/security/jwt/model/TokenValidationRequest.java
@@ -1,14 +1,0 @@
-package user.security.jwt.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenValidationRequest {
-
-    private TokenDto tokenDto;
-
-}
