--- user/src/main/java/user/security/jwt/model/TokenValidationResponse.java
@@ -1,16 +1,0 @@
-package user.security.jwt.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TokenValidationResponse {
-
-    private Long userId;
-
-}
